Server can now work without a secret
The previous modification made mandatory the secret parameter, that
means if the server was running in a contained and secure network a
shared secret would have to be generated and configured.

We are allowing now that a secret is not defined and in those cases
we are not checking for any secret sent in the request.